Test: all tab cases are working!
