+++ test/test_lib/test_case.hpp
@@ -306,8 +306,7 @@ typedef enum {
     JSON_READ = (1<<4),
     HAS_CONTAINER_KEYS = (1<<5),
     HAS_MULTILINE_SCALAR = (1<<6),
-    NO_TREE_PARSE = (1<<7),
-    NO_COMPARE_EMITTED = (1<<8),
+    NO_COMPARE_EMITTED = (1<<7),
 } TestCaseFlags_e;
